requirements.txt
python-dotenv
pymongo~=4.4.1
Flask[async]>=2.1.1
pandas
gunicorn
